Check for conversation and create new one if not yet created
